Address review feedback on helpers, allowlist, and popover dismissal
