+++ test/components/ansi-log-chunks.test.ts
@@ -1,10 +1,5 @@
-/**
- * @vitest-environment happy-dom
- */
-// ansi-log transitively imports the wa-icon component (for the doc-link
-// annotations), which needs DOM globals (CSSStyleSheet) at module load.
 import { describe, expect, it } from "vitest";
-import { chunksToVisualLines } from "../../src/components/ansi-log.js";
+import { chunksToVisualLines } from "../../src/util/log-chunks.js";
 
 describe("chunksToVisualLines", () => {
   it("appends \\n-terminated chunks as discrete lines", () => {
